requirements.txt
pycryptodome
psycopg2-binary
pika
pyyaml